requirements.txt
httpx
pandas
sentence-transformers
reportlab
python-docx
pypdf2
fastapi
pydantic
requests
scikit-learn
streamlit